+++ app/src/main/java/org/breezyweather/ui/main/dialogs/ErrorHelpDialog.kt
@@ -37,7 +37,8 @@ object ErrorHelpDialog {
         activity: Activity,
         @StringRes title: Int,
         @StringRes content: Int,
-        showSettings: Boolean,
+        @StringRes action: Int? = null,
+        onActionClick: ((Activity) -> Unit)? = null,
     ) {
         val view = LayoutInflater
             .from(activity)
@@ -71,15 +72,15 @@ object ErrorHelpDialog {
                                 Text(stringResource(id = R.string.action_close))
                             }
                         },
-                        dismissButton = if (showSettings) {
+                        dismissButton = if (action != null && onActionClick != null) {
                             {
                                 TextButton(
                                     onClick = {
-                                        IntentHelper.startWeatherProviderSettingsActivity(activity)
+                                        onActionClick(activity)
                                         dialogOpenState.value = false
                                     }
                                 ) {
-                                    Text(stringResource(id = R.string.action_settings))
+                                    Text(stringResource(id = action))
                                 }
                             }
                         } else {